Move loaders to CSET.loaders
