+++ src/CSET/loaders/histograms.py
