+++ src/CSET/loaders/spatial_difference_field.py
