+++ src/CSET/loaders/spatial_field.py
